% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.data.R
\name{download.data}
\alias{download.data}
\title{Download data for TRIM}
\usage{
download.data(download.path = NULL)
}
\arguments{
\item{download.path}{Folder to store downloaded data. By default the data is stored at "data/" in the package folder}
}
\description{
Download data for TRIM
}
