% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_functions.R
\name{simple.trim}
\alias{simple.trim}
\title{simple.trim}
\usage{
simple.trim(Y.obs, A, trim.side = NULL, alpha = NULL)
}
\arguments{
\item{Y.obs}{a numeric vector of clinical endpoints which includes \code{NA} values.}

\item{A}{a numeric vector containing a binary indicator of treatment assingment.}

\item{trim.side}{a character string specifying the side of the distribution for the clinical endpoint corresponding to a poor outcome that will be trimmed in the analysis, must be one of "upper" or "lower".}

\item{alpha}{a number indicating the fraction of data to be trimmed from each arm before calculating the trimmed means. This fraction must be less than alpha and greater than or equal to the maximum fraction of missing data between treatment groups. If \code{NULL} then alpha will be automatically set to the maximum fraction of missing data between treatment groups.}
}
\value{
A vector containing the trimmed means of each treatment group and the difference between the trimmed means
}
\description{
Calculates trimmed means and differences for two treatment groups after ranking missing values to the poor end of the distribution.
}
\examples{
# Adaptive alpha
simple.trim(Y.obs = c(rnorm(50),rnorm(50,-1)),A = rep(c(0,1),each=50),trim.side="upper",alpha=NULL)

# Fixed alpha = 0.5
simple.trim(Y.obs = c(rnorm(50),rnorm(50,-1)),A = rep(c(0,1),each=50),trim.side="upper",alpha=0.5)
}
