\name{trim-package}
\alias{trim-package}
\alias{trim}
\docType{package}
\title{
Tree methods for classification of imbalanced data
}
\description{
This package provides ad hoc tree methods for dealing with classification of imbalanced variables.
}
\details{
\tabular{ll}{
Package: \tab trim\cr
Type: \tab Package\cr
Version: \tab 0.1-1\cr
Date: \tab 2013-05-10\cr
License: \tab GPL-3\cr
}
~~ An overview of how to use the package, ~~
~~ including the most important functions ~~
}
\author{
Emmanuel Rousseaux

Maintainer: Emmanuel Rousseaux <emmanuel.rousseaux@unige.ch>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\keyword{ trim }
\keyword{ tree }
\keyword{ classification }
\keyword{ imbalanced }
\keyword{ unbalanced }
\keyword{ data }

\seealso{
Tree for global optimization: \code{\link[evtree:evtree-package]{evtree}}.
}
\examples{

}
