% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_functions.R
\name{trim}
\alias{trim}
\title{trim}
\usage{
trim(Y.obs, A, trim.side = NULL, alpha = NULL, num.perm = 10000,
  seed = 907)
}
\arguments{
\item{Y.obs}{a numeric vector of clinical endpoints which includes some \code{NA} values.}

\item{A}{a numeric vector containing a binary indicator of treatment assingment.}

\item{trim.side}{a character string specifying the side of the distribution for the clinical endpoint corresponding to a poor outcome that will be trimmed in the analysis, must one of "upper" or "lower". The default is \code{NULL} which throws an error: 'trim.side argument not specified'.}

\item{alpha}{a number indicating the fraction of data to be trimmed from each arm before calculating the trimmed means. This fraction must be less than alpha and greater than or equal to the maximum fraction of missing data between treatment groups. If \code{NULL} then alpha will be automatically set to the maximum fraction of missing data between treatment groups.}

\item{num.perm}{a number specifying how many permutations to use when generating the permutation distribution used to build confidence intervals and calculate a p-value. Defaults to \code{10000}.}

\item{quantile.CI}{a logical argument indicating }


\value{
A matrix containing the trimmed means of each treatment group, the difference between the trimmed means, confidence interval, and p-value.
}
\description{
Estimates treatment effects for a randomized clinical trial data using the trimmed means approach for missing data introduced by Permutt & Li (2017). This function calculates trimmed means and differences for two treatment groups after ranking missing values to the poor end of the distribution. Inference is conducted via a permutation test.
}
\examples{
# adaptive alpha (Note: default alpha = NULL)
trim(Y.obs = c(rnorm(50),rnorm(50,-1)),A = rep(c(0,1),each=50),trim.side="upper")

# fixed alpha = 0.5, increased permutations, seed not set
trim(Y.obs = c(rnorm(50),rnorm(50,-1)),A = rep(c(0,1),each=50),trim.side="upper",alpha=0.5,num.perm=30000,seed=NULL)
}
\references{
Permutt, Thomas, and Feng Li. "Trimmed means for symptom trials with dropouts." Pharmaceutical statistics 16.1 (2017): 20-28.
}
